Basic inventory menu
The inventory menu now shows! You still can't do anything with it, though.
The position of the "buttons" have been changed to be aligned with the rest of the text.
There is also a "checkItem" boolean that determines if the game has to check for anything, but it also does nothing right now.