Separated world setup from __init__
As it turns out a function won't work, but I can just write the plain code inside the module.
So that's where the world code is going to stay for now. I'll most likely be editing that a lot, for when I set up new locations and such.
All that's left to do before 0.1.0 is to allow people to use items and set up a slightly bigger world.